architecture finished

+++ vote_sphere/lib/infrastructure/repositories/settings_repository.dart
@@ -1,41 +1,17 @@
-import 'dart:async';
 import 'dart:convert';
-import 'dart:io';
-
-import 'package:bloc/bloc.dart';
-import 'package:meta/meta.dart';
-import 'package:vote_sphere/local_storage/secure_storage.dart';
 import 'package:http/http.dart' as http;
+import 'package:vote_sphere/infrastructure/local_storage/secure_storage.dart';
 
-part 'settings_event.dart';
-part 'settings_state.dart';
-
-class SettingsBloc extends Bloc<SettingsEvent, SettingsState> {
-  SettingsBloc() : super(SettingsInitial()) {
-    on<LoadSettingEvent>(loadSettingEvent);
-    on<NavigateToChangePasswordEvent>(navigateToChangePasswordEvent);
-    on<ChangePaswordEvent>(changePaswordEvent);
-    on<DeleteAccountEvent>(deleteAccountEvent);
-  }
-
-  FutureOr<void> loadSettingEvent(
-      LoadSettingEvent event, Emitter<SettingsState> emit) async {
-    emit(SettingsLoadingState());
-
+class SettingsRepository {
+  static Future<Map> loadSetting(event) async {
     final secureStorage = SecureStorage().secureStorage;
     final username = await secureStorage.read(key: 'username');
     final email = await secureStorage.read(key: "email");
 
-    emit(SettingsLoadedState(username: username, email: email));
+    return {"username": username, "email": email};
   }
 
-  FutureOr<void> navigateToChangePasswordEvent(
-      NavigateToChangePasswordEvent event, Emitter<SettingsState> emit) {
-    emit(NavigateToUpdatePasswordState());
-  }
-
-  FutureOr<void> changePaswordEvent(
-      ChangePaswordEvent event, Emitter<SettingsState> emit) async {
+  static Future<String> changePassword(event) async {
     String uri = 'http://localhost:9000/auth/changePassword';
     final url = Uri.parse(uri);
     final secureStorage = SecureStorage().secureStorage;
@@ -48,21 +24,20 @@ class SettingsBloc extends Bloc<SettingsEvent, SettingsState> {
     };
 
     final res = await http.patch(url, headers: headers, body: jsonBody);
-    print(res.body);
+    print(res.statusCode);
     if (res.statusCode == 200) {
-      emit(ChangePasswordSuccessState());
+      return "success";
     } else {
       final decodedBody = jsonDecode(res.body);
       var error = decodedBody["message"];
       if (decodedBody["message"] is! String) {
         error = decodedBody["message"][0];
       }
-      emit(ChangePasswordErrorState(error: error));
+      return error;
     }
   }
 
-  FutureOr<void> deleteAccountEvent(
-      DeleteAccountEvent event, Emitter<SettingsState> emit) async {
+  static Future<bool> deleteAccount() async {
     final secureStorage = SecureStorage().secureStorage;
     final token = await secureStorage.read(key: 'token');
     String uri = 'http://localhost:9000/auth/deleteAccount';
@@ -72,7 +47,11 @@ class SettingsBloc extends Bloc<SettingsEvent, SettingsState> {
       'Authorization': 'Bearer $token'
     };
     final res = await http.delete(url, headers: headers);
-    print(res.body);
-    await secureStorage.deleteAll();
+    if (res.statusCode >= 200 && res.statusCode < 300) {
+      await secureStorage.deleteAll();
+      return true;
+    } else {
+      return false;
+    }
   }
 }
